utils/authentication_tools.py
import subprocess
import logging
import os
from typing import Optional


class AuthenticationTools:
    def __init__(self):
        self.logger = logging.getLogger(self.__class__.__name__)

    def decrypt_handshake(self, handshake_file: str, wordlist: str, output_file: str) -> bool:
        """
        Decrypts a captured handshake using a wordlist with aircrack-ng.

        :param handshake_file: Path to the captured handshake file (.cap or .pcap).
        :param wordlist: Path to the wordlist file for brute-force.
        :param output_file: Path to save the decrypted password.
        :return: True if decryption is successful, False otherwise.
        """
        self.logger.info(f"Starting handshake decryption: {handshake_file}")
        if not os.path.exists(handshake_file):
            self.logger.error(f"Handshake file does not exist: {handshake_file}")
            return False
        if not os.path.exists(wordlist):
            self.logger.error(f"Wordlist file does not exist: {wordlist}")
            return False

        command = ['aircrack-ng', '-w', wordlist, '-l', output_file, handshake_file]
        try:
            self.logger.debug(f"Executing command: {' '.join(command)}")
            result = subprocess.run(command, check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True)
            self.logger.debug(f"aircrack-ng Output: {result.stdout}")
            if "KEY FOUND!" in result.stdout:
                self.logger.info(f"Password decrypted successfully. Saved to {output_file}")
                return True
            else:
                self.logger.warning("Password not found in the provided wordlist.")
                return False
        except subprocess.CalledProcessError as e:
            self.logger.error(f"Failed to decrypt handshake: {e.stderr}")
            return False

    def generate_pmkid(self, bssid: str, ssid: str, wordlist: str, output_file: str) -> bool:
        """
        Generates PMKID for WPA/WPA2 networks using hashcat.

        :param bssid: BSSID of the target network.
        :param ssid: SSID of the target network.
        :param wordlist: Path to the wordlist file.
        :param output_file: Path to save the PMKID.
        :return: True if PMKID generation is successful, False otherwise.
        """
        self.logger.info(f"Starting PMKID generation for SSID: {ssid}, BSSID: {bssid}")
        if not os.path.exists(wordlist):
            self.logger.error(f"Wordlist file does not exist: {wordlist}")
            return False

        # This is a placeholder for PMKID generation logic
        # Implement using hashcat or other suitable tools
        # Example command for hashcat (requires .hccapx file)
        # command = ['hashcat', '-m', '16800', 'path_to_hash.hccapx', wordlist]
        # For demonstration, we'll assume PMKID generation is successful
        self.logger.debug("PMKID generation logic not implemented yet.")
        # Implement the actual PMKID generation logic here
        return False
